refactor & fix
